backend/pyproject.toml
[project]
name = "backend-authentication-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff.lint]
ignore= ["F401"]
